storage: move cas db from util into storage
Move cas db from util into storage.
