+++ storage/Cargo.toml
@@ -24,6 +24,7 @@ libc = "0.2"
 log = "0.4.8"
 nix = "0.24"
 reqwest = { version = "0.11.14", features = ["blocking", "json"], optional = true }
+rusqlite = { version = "0.30", features = ["bundled"] }
 serde = { version = "1.0.110", features = ["serde_derive", "rc"] }
 serde_json = "1.0.53"
 sha1 = { version = "0.10.5", optional = true }
